enc: guard hairpin track range and harden against malformed input
Add the validTrack guard the ottava resolver already uses to the hairpin
resolver, so a grand-staff wedge whose note carries an out-of-range voice
nibble no longer builds a Hairpin at a track beyond ntracks and asserts at
layout. Add hostile-fixture no-crash tests: crafted zero-tuplet-nibble,
out-of-range staff/voice and zero-size-element inputs, plus truncation at
every block boundary and across the v0xA6 absolute-seek header.

Also widen computeElementDurations durTicks to int, drop the redundant
v0xC2 readInstrumentMeta override and the NOTE/REST xoffset type-switch
casts (one inherited base field), and remove dead members EncMeasure::m_id,
EncInstrument::showStaff and the EncAccidentalType enum.

+++ src/importexport/encore/internal/parser/parsers-measure.cpp
@@ -87,7 +87,7 @@ static std::unique_ptr<EncMeasureElem> createMeasureElement(
 // mid-measure clef change is not stretched to fill the whole measure.
 void computeElementDurations(
     std::vector<EncMeasureElem*>& elems,
-    qint16 durTicks,
+    int durTicks,
     bool hasGraceTimeBorrowing,
     const std::vector<qint16>& boundaryTicks)
 {
